+++ library/src/validations/mac/mac.ts
@@ -1,6 +1,6 @@
 import { MAC48_REGEX, MAC64_REGEX } from '../../regex.ts';
 import type { BaseValidation, ErrorMessage } from '../../types/index.ts';
-import { getOutput, getPipeIssues } from '../../utils/index.ts';
+import { actionIssue, actionOutput } from '../../utils/index.ts';
 
 /**
  * MAC validation type.
@@ -17,7 +17,7 @@ export type MacValidation<TInput extends string> = BaseValidation<TInput> & {
 };
 
 /**
- * Creates a validation function that validates a [MAC address](https://en.wikipedia.org/wiki/MAC_address).
+ * Creates a validation function that validates a [MAC](https://en.wikipedia.org/wiki/MAC_address).
  *
  * @param message The error message.
  *
@@ -34,8 +34,8 @@ export function mac<TInput extends string>(
     _parse(input) {
       return !this.requirement[0].test(input) &&
         !this.requirement[1].test(input)
-        ? getPipeIssues(this.type, this.message, input, this.requirement)
-        : getOutput(input);
+        ? actionIssue(this.type, this.message, input, this.requirement)
+        : actionOutput(input);
     },
   };
 }